Added some extra stuff for the power_board file, ioncluding a heartbeat where we send the voltage reading every 10 seconds.
