+++ Core/Inc/power_board/power_board.h
@@ -5,6 +5,14 @@
 #include "peripheral_util.h"
 #include "CanDriver.h"
 
+
+/*CAN BUS flags*/
+extern uint16_t heartbeat_10000ms;
+bool kill_flag;
+bool voltage_request;
+uint16_t voltage_reading;
+uint64_t TxMailbox[0];  // Declaration of an array used in CAN
+
 void init();
 void loop();
